requirements.txt
flask
matplotlib<4.0
numpy<1.25
pandas<2.0
python-dotenv